Move Deduper into a separate file
