+++ perf/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod cuda_runtime;
 pub mod data_budget;
+pub mod deduper;
 pub mod discard;
 pub mod packet;
 pub mod perf_libs;
